Split UserNotification into provider/controller
