+++ lib/Alchemy/Phrasea/Application.php
@@ -14,7 +14,6 @@ namespace Alchemy\Phrasea;
 use Alchemy\Geonames\GeonamesServiceProvider;
 use Alchemy\Phrasea\ControllerProvider\Thesaurus\Thesaurus;
 use Alchemy\Phrasea\ControllerProvider\Thesaurus\Xmlhttp as ThesaurusXMLHttp;
-use Alchemy\Phrasea\ControllerProvider\User\Notifications;
 use Alchemy\Phrasea\ControllerProvider\User\Preferences;
 use Alchemy\Phrasea\Core\Event\Subscriber\BasketSubscriber;
 use Alchemy\Phrasea\Core\Event\Subscriber\BridgeSubscriber;
@@ -317,6 +316,7 @@ class Application extends SilexApplication
             'Alchemy\Phrasea\ControllerProvider\Root\Root' => [],
             'Alchemy\Phrasea\ControllerProvider\Root\RSSFeeds' => [],
             'Alchemy\Phrasea\ControllerProvider\Root\Session' => [],
+            'Alchemy\Phrasea\ControllerProvider\User\Notifications' => [],
             'Alchemy\Phrasea\ControllerProvider\Datafiles' => [],
             'Alchemy\Phrasea\ControllerProvider\Lightbox' => [],
             'Alchemy\Phrasea\ControllerProvider\MediaAccessor' => [],
@@ -622,8 +622,6 @@ class Application extends SilexApplication
     public function bindRoutes()
     {
         $this->mount('/user/preferences/', new Preferences());
-        $this->mount('/user/notifications/', new Notifications());
-
 
         $this->mount('/thesaurus', new Thesaurus());
         $this->mount('/xmlhttp', new ThesaurusXMLHttp());
@@ -681,6 +679,7 @@ class Application extends SilexApplication
             '/report/'                     => 'Alchemy\Phrasea\ControllerProvider\Report\Root',
             '/session/'                    => 'Alchemy\Phrasea\ControllerProvider\Root\Session',
             '/setup'                       => 'Alchemy\Phrasea\ControllerProvider\Setup',
+            '/user/notifications/'         => 'Alchemy\Phrasea\ControllerProvider\User\Notifications',
             '/'                            => 'Alchemy\Phrasea\ControllerProvider\Root\Root',
         ];
